requirements.txt
# Core IoT & MQTT
paho-mqtt==1.6.1
asyncio-mqtt==0.16.1

# Firebase
firebase-admin==6.2.0
google-cloud-firestore==2.12.0

# Data Processing & ML
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.0

# Web Server
flask==2.3.2
flask-cors==4.0.0
flask-restful==0.3.10
requests==2.31.0

# Database
sqlalchemy==2.0.19
pymongo==4.4.0

# Time Series & Forecasting
statsmodels==0.14.0
prophet==1.1.4

# Utilities
python-dotenv==1.0.0
pyyaml==6.0
logging-config==0.5.0

# API & Serialization
jsonschema==4.18.2
marshmallow==3.19.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.0.0
pylint==2.17.5
